update project structure (rename backend + replace obstacles)
