+++ Backend/CMakeLists.txt
@@ -1,8 +1,7 @@
 cmake_minimum_required(VERSION 3.20)
 
-project(AdjacencyMatrix VERSION 0.1 LANGUAGES CXX)
+project(backend VERSION 0.1 LANGUAGES CXX)
 
 file(GLOB PROJECT_SOURCES "*.h" "*.cpp")
 
-add_library(AdjacencyMatrix STATIC ${PROJECT_SOURCES})
-
+add_library(backend STATIC ${PROJECT_SOURCES})
