+++ gui/circle_obstacle.h
