+++ gui/poly_obstacle.h
