+++ lib/poly_obstacle.h
